Byttet ut oppslag mot dagpenger-oppslag med brreg enhetsregister navik/dagpenger#539
