+++ .github/workflows/deploy.yaml
@@ -38,7 +38,7 @@ jobs:
   deploy-preprod:
     name: Deploy to Preprod
     needs: build
-    if: github.ref == 'refs/heads/master'
+    if: github.ref == 'refs/heads/task/fjerne-dagpenger-oppslag'
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@master
